Floating point end-time
git-svn-id: file:///home/amuse/svn/amuse/trunk@4824 6e832d89-e7d1-4ad1-8ca1-873beecccf65

+++ src/amuse/community/bonsai/src/include/octree.h
@@ -284,7 +284,7 @@ protected:
   int   dt_limit;
   float eta;
   float timeStep;
-  int   tEnd;
+  float   tEnd;
   float theta;
 
   //Sim stats
